requirements.txt
coverage==4.5.4
Django==2.2.4
djangorestframework==3.10.3
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
